+++ resources/lang/fr/texts.php
@@ -1,6 +1,6 @@
 <?php
 
 return [
-    'message' => 'Ce site nécessite l\'autorisation de cookies pour fonctionner correctement.',
+    'message' => "Ce site nécessite l'autorisation de cookies pour fonctionner correctement.",
     'agree' => 'Accepter',
 ];