+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-All Notable changes to `laravel-cookie-consent` will be documented in this file
+All notable changes to `cookie-consent` will be documented in this file
+
+## 2.0.0 - 2017-XX-XX
+
+- add support for Laravel 5.5, drop support for older versions of the framework.
+- renamed config file from `laravel-cookie-consent` to `cookie-consent`.
 
 ## 1.9.0 - 2017-08-22